refactor: Separate http probe component
